Fix git lowercase bug, step 1
