--- src/main/java/com/sonicether/soundphysics/SPGUIFactory.java
@@ -1,40 +1,0 @@
-package com.sonicether.soundphysics;
-
-import java.util.Set;
-
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.GuiScreen;
-import net.minecraftforge.fml.client.IModGuiFactory;
-
-public class SPGuiFactory implements IModGuiFactory {
-
-	@Override
-	public void initialize(Minecraft minecraftInstance) {
-	}
-
-	@Override
-	public boolean hasConfigGui() {
-		return true;
-	}
-
-	@Override
-	public GuiScreen createConfigGui(GuiScreen parentScreen) {
-		return new SPGuiConfig(parentScreen);
-	}
-
-	@Override
-	public Class<? extends GuiScreen> mainConfigGuiClass() {
-		return SPGuiConfig.class;
-	}
-
-	@Override
-	public Set<RuntimeOptionCategoryElement> runtimeGuiCategories() {
-		return null;
-	}
-
-	@Override
-	public RuntimeOptionGuiHandler getHandlerFor(RuntimeOptionCategoryElement element) {
-		return null;
-	}
-
-}
